[Issue #108]: replace hdfs FileSystem api with the unified Storage api. (#110)
With the unified Storage interface, we can easily add different types of storage systems (e.g. S3, local fs, Ozone...) into Pixels.

In this PR, we have already implemented the HDFS storage, which internally use APIs in hadoop libraries to access HDFS.
The local file storage is still to be implemented.

For this PR, as we add a new column into the TBLS table in the metadata database, some components like pixels-presto may be not compatible with a component from the previous version.