[Issue #172]: implement record cursor and enhance record reader. (#173)
1. Add the PixelsRecordSet and PixelsRecordCusor, which can be enabled/disabled by configuration properties, to pixels-presto.
2. Update PixelsRecordReaderImpl.readBatch, let users specify whether they want to reuse the returned result row batch.

Issue #172 is not completely resolved. Record cursor is much slower than the page source for wide tables (they have similar performance on TPC-H). We also tried to implement a context (buffer) of result row batches to automatically allocate, free, and reuse the result row batches. However, it reduces the query performance a lot. This issue can be solved by implementing the read path in C++ in the future.