[Issue #222]: disable mock file locations for storage systems that do not provide data locality. (#223)
For storage systems such as S3, which does not provide physical data locations, we do not generate mock file locations using etcd. This optimization helps improve the scalability of Pixels. It is important for scanning data using cloud functions as there might be a large number of CF instances scanning data in parallel.