[Issue #419] fix timezone offset for date column. (#422)
During data writing, we read a date in string format, such as '1999-12-31', parse it to a date object in local time, and get the elapsed milliseconds since the Unix epoch '1970-01-01 00:00:00 UTC'. However, SQL query engines such as Trino accepts the epoch days for date type. If we parsed the string input using local time, we must add the timezone offset to the epoch milliseconds of local time before deciding it by 86400000 (milliseconds in a day).