[Issue #498] remove the orders array from dictionary encoding. (#499)
Previously, the orders array was the mapping from the encoded id to the key index in the dictionary. We remove the orders array from the pixels file by enforcing the dictionary to be written in the order of encoded id.
This helps improve writing and reading performance, as it reduces the CPU costs on array encoding/decoding and the IO costs.