[Issue #549] fix dependencies for slf4j (#550)
Add log4j-slf4j-impl as the default binding of slf4j.
Add a unified dependency definition for slf4j-api and exclude slf4j-api dependency in other libs including jetcd, jedis, and hadoop. But we do not add this dependency to the modules by default. It can be added as needed in the future.