[Issue #996] move the partitioned cache implementation to the legacy package (#997)
The old partitioned cache was implemented by @Yeeef.
This implementation is incomplete, primarily exploring and experimenting with different cache indexing solutions and optimizations. This implementation is never used in pixels-cache, so it has been moved to the legacy package. We may use insights gained from this implementation to optimize the pixels-cache implementation in the future.